(fix): likely overflow error

+++ tests/test_preprocessing.py
@@ -406,7 +406,9 @@ def test_regress_out_ordinal():
 def test_regress_out_layer(dtype):
     from scipy.sparse import random
 
-    adata = AnnData(random(1000, 100, density=0.6, format="csr", dtype=dtype))
+    adata = AnnData(
+        random(1000, 100, density=0.6, format="csr", dtype=np.uint16).astype(dtype)
+    )
     adata.obs["percent_mito"] = np.random.rand(adata.X.shape[0])
     adata.obs["n_counts"] = adata.X.sum(axis=1)
     if dtype == np.float64:
